Add subject templates management to class sections page
