+++ app/src/pages/ClassSections/components/index.ts
@@ -2,4 +2,5 @@ export { ClassSectionHeader } from './ClassSectionHeader'
 export { ClassSectionList } from './ClassSectionList'
 export { ClassSectionModal } from './ClassSectionModal'
 export { ClassSectionSubjects } from './ClassSectionSubjects'
-export { ClassSectionSubjectModal } from './ClassSectionSubjectModal' 
+export { ClassSectionSubjectModal } from './ClassSectionSubjectModal'
+export { SubjectTemplatesModal } from './SubjectTemplatesModal'
